change log:
+ added tank object
+ added bullet object (still haven't fixed that bug when you delete
bullets)

+++ Java2DGame/src/utils/Vector2.java
@@ -1,10 +1,10 @@
 package utils;
 
 public class Vector2 {
-	public float x;
-	public float y;
+	public double x;
+	public double y;
 	
-	public Vector2(float x, float y) {
+	public Vector2(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
